pyproject.toml
[project]
name = "ai-tutor"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.12",
    "google-adk>=1.0.0",
    "numpy>=2.2.6",
    "plotly>=6.1.1",
    "schemdraw>=0.20",
    "sympy>=1.14.0",
]
